RESET MIGRATIONS
- Change PlayerCharacter* to use an auto-generated primary key, keying anything by Blizzard Id is a mistake
- Add PlayerCharacterReputations and move data to it
- Add Team/TeamCharacter models

+++ src/Wowthing.Lib/Models/Query/SchedulerCharacterQuery.cs
@@ -10,7 +10,7 @@ namespace Wowthing.Lib.Models.Query
         public long UserId { get; set; }
         public WowRegion Region { get; set; }
         public string RealmSlug { get; set; }
-        public long CharacterId { get; set; }
+        public int CharacterId { get; set; }
         public string CharacterName { get; set; }
     }
 }
